MLP inference +  DuckDB search +,Pytorch Regressor
