+++ .gitignore
@@ -199,3 +199,4 @@ cython_debug/
 bash_scripts/jdk-25_linux-x64_bin.tar
 bash_scripts/jdk-17.0.12_linux-x64_bin.tar.gz
 bash_scripts/jdk-25_linux-x64_bin.tar.gz
+uv.lock
